# services/api/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Shipped API
  version: "1.0.0-phase2"
  description: >
    The Shipped control plane. Most /v1 routes require a Bearer EdDSA JWT
    (issued by Better Auth) which the Go API verifies itself (the authz boundary).
    Every request runs under a Postgres RLS tenant context derived from the
    verified org_id/user_id claims.


    Phase 2 adds access control & domains: the /authz viewer exchange (mint +
    password) which issues a host-scoped EdDSA edge token the serving Worker
    verifies (public JWKS at /.well-known/edge-jwks), per-site access-mode + policy
    management (admin/owner only), allowlist CRUD, the org allow_external_sharing
    toggle (with reconcile), member listing, and Cloudflare-for-SaaS custom domains.
  license:
    name: FSL-1.1-Apache-2.0
servers:
  - url: https://api.shipped.app
    description: Production
  - url: http://localhost:8080
    description: Local / self-host
security:
  - bearerAuth: []
paths:
...
  /.well-known/edge-jwks:
    get:
      summary: Edge-token JWKS (unauthenticated)
      description: >
        The public JWKS (OKP/Ed25519) for the edge-token signer. The serving Worker
        fetches this and pins alg=EdDSA to verify the host-scoped edge token. This
        is a SEPARATE keypair from Better Auth's user JWKS.
      security: []
      operationId: edgeJwks
      responses:
        "200":
          description: The edge signer's public JWKS
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JWKS" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/authz/mint:
    post:
      summary: Mint a host-scoped edge token for an org_only/allowlist site
      description: >
        Authorizes the VERIFIED viewer (Better Auth JWT) for the org_only/allowlist
        site resolved from {host} and mints a host-scoped EdDSA edge token (sub =
        viewer). Re-checks live tables: org membership for org_only; the viewer's
        VERIFIED email on the allowlist for allowlist (claiming a pending entry on
        first match; external grants require allow_external_sharing). An expired
        policy refuses the mint (403). A password-mode site → 400 (use /authz/password).
      operationId: authzMint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host]
              properties:
                host: { type: string, description: Content host (e.g. acme.shippedusercontent.com or a verified custom host) }
                next: { type: string, description: Path the dashboard redirects back to (NOT used for authz) }
      responses:
        "200":
          description: Minted edge token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EdgeToken" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/authz/password:
    post:
      summary: Mint an anonymous edge token for a password-protected site
      description: >
        JWT-FREE. Verifies the submitted password against the site's bcrypt hash
        (constant-time compare) and, on success, mints an ANONYMOUS host-scoped edge
        token (sub = "anon:<random>"). Wrong password → 401; an expired policy →
        403; a non-password / unknown host → a generic 401 (no existence oracle).
      security: []
      operationId: authzPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host, password]
              properties:
                host: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Minted anonymous edge token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EdgeToken" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /v1/domains/{domainID}/status:
    get:
      summary: Poll a custom domain's verification status
      description: >
        Polls Cloudflare for the custom hostname's verification + TLS state, advances
        the pending→verifying→active state machine, and — on verified+TLS — writes
        the global host route → site so the custom host serves.
      operationId: getDomainStatus
      parameters:
        - name: domainID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The updated domain row
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /v1/sites/{id}:
    get:
      summary: Get one site
      operationId: getSite
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      responses:
        "200":
          description: The site
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/sites/{id}/domains:
    get:
      summary: List a site's custom domains
      operationId: listDomains
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      responses:
        "200":
          description: The site's custom domains
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Register a custom domain for a site (admin/owner only)
      description: >
        Owner/admin only. Creates a Cloudflare-for-SaaS custom hostname and stores a
        pending domains row, returning the DNS DCV record the user must create.
        hostname is GLOBALLY UNIQUE → 409 if taken by another org/site.
      operationId: addDomain
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname]
              properties:
                hostname: { type: string, example: docs.acme.com }
      responses:
        "201":
          description: Custom domain registered (pending verification)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived EdDSA JWT (Better Auth) or a deploy token.
  parameters:
    SiteID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: app.sites.id
  schemas:
...
    EdgeToken:
      type: object
      description: A minted host-scoped EdDSA edge token (the serving Worker verifies it).
      properties:
        token: { type: string, description: Compact EdDSA JWT (aud = the content host). }
        host: { type: string }
        mode: { type: string, enum: [password, allowlist, org_only] }
...
    Domain:
      type: object
      properties:
        id: { type: string, format: uuid }
        site_id: { type: string, format: uuid }
        hostname: { type: string }
        verify_status: { type: string, enum: [pending, verifying, verified, failed] }
        tls_status: { type: string, enum: [pending, issued, failed] }
        dcv_record: { type: string, description: "DNS record to create: '<name> <type> <value>'." }
    JWKS:
      type: object
      description: JSON Web Key Set (OKP/Ed25519) for the edge-token signer.
      properties:
        keys:
          type: array
          items:
            type: object
            properties:
              kty: { type: string, example: OKP }
              crv: { type: string, example: Ed25519 }
              kid: { type: string }
              x: { type: string, description: base64url(raw public key) }
              use: { type: string, example: sig }
              alg: { type: string, example: EdDSA }
    Site:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        slug: { type: string }
        owner_id: { type: string, format: uuid }
        access_mode: { type: string, enum: [public, password, allowlist, org_only] }
        current_version_id: { type: string, format: uuid, nullable: true }
        live_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string, description: Stable machine-readable code }
        message: { type: string }
...
  responses:
    Unauthorized:
      description: Missing/invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted (e.g. not admin/owner, expired link, not on allowlist, external sharing disabled)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unavailable:
      description: A required dependency (DB, edge signer, custom-domain provider) is not configured
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Malformed request / reserved slug / verification failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource absent (or invisible under the caller's tenant)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Slug/host already in use by another org/site (global host registry)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...